requirements-test.txt
# unit + integration + api
pytest
pytest-cov
pytest-flask

# e2e
pytest-playwright

# After installing, run: playwright install chromium
